Given an array of indices find the number of triplets (i,j,k) such that the elements of those indices are in geometric progression for a given common ratio r and i < j < k.
Was trying to solve this on four different days but finally managed it. Idea is to keep track of a middle element arr[i] which is divisible
by r and then finding out how many elements to the left have value arr[i]/r and how many to the right have value arr[i]*r. Use two
dictionaries to keep count for each.